Add CenteredMap reusable component
